requirements.txt
yt-dlp
openai-whisper
torch
deep-translator
langdetect
tqdm
numpy